refactor: replace existing useLogout hook with useLogoutFlow hook
