+++ packages/hash/frontend/src/shared/layout/layout-with-sidebar/workspace-switcher.tsx
@@ -16,9 +16,9 @@ import {
 } from "material-ui-popup-state/hooks";
 import { Avatar, FontAwesomeIcon } from "@hashintel/hash-design-system";
 import { useUser } from "../../../components/hooks/useUser";
-import { useLogout } from "../../../components/hooks/useLogout";
 import { Button, MenuItem } from "../../ui";
 import { useRouteAccountInfo } from "../../routing";
+import { useLogoutFlow } from "../../../components/hooks/useLogoutFlow";
 
 type WorkspaceSwitcherProps = {};
 
@@ -29,7 +29,7 @@ export const WorkspaceSwitcher: VFC<WorkspaceSwitcherProps> = () => {
     popupId: "workspace-switcher-menu",
   });
   const { user } = useUser();
-  const { logout } = useLogout();
+  const { logout } = useLogoutFlow();
   const { accountId } = useRouteAccountInfo();
 
   const activeWorkspace = useMemo(() => {
